🎨 팟 페이지 퍼블리싱 리팩토링
