+++ src/app/(with-main-header)/party/create/CreatePartyButton.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Button from "@/components/common/Button";
 import InputWithErrorMsg from "@/components/common/InputWithErrorMsg";
 import Modal from "@/components/common/Modal";
@@ -7,8 +5,8 @@ import useInputValidation from "@/hooks/useInputValidation";
 import { useModalStore } from "@/store/modalStore";
 import { FormEvent } from "react";
 
-export default function ParticipateButton() {
-  const { open } = useModalStore((state) => state);
+export default function CreatePartyButton() {
+  const { open, close } = useModalStore((state) => state);
   const { validate, ...point } = useInputValidation(0, (value) => {
     if (!value || Number(value) < 100) {
       return "최소 배팅 금액 이상 입력해주세요";
@@ -16,7 +14,7 @@ export default function ParticipateButton() {
     return null;
   });
 
-  const participateParty = (e: FormEvent<HTMLFormElement>) => {
+  const createParty = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
 
     if (validate()) {
@@ -24,28 +22,28 @@ export default function ParticipateButton() {
       close();
     }
   };
+
   return (
     <>
-      <div className="flex justify-end h-15">
-        <Button onClick={() => open("participate-party")}>참여하기</Button>
-      </div>
-
-      <Modal modalId="participate-party">
+      <Button type="button" onClick={() => open("create-party")}>
+        생성하기
+      </Button>
+      <Modal modalId="create-party">
         <form
-          onSubmit={participateParty}
+          onSubmit={createParty}
           className="flex flex-col items-center w-full gap-7.5"
         >
           <div className="flex flex-col w-full items-center gap-4">
             <span className="text-xl">배팅 포인트</span>
             <InputWithErrorMsg
+              autoFocus
               type="number"
               className="bg-site-button-input"
               placeholder="배팅할 포인트를 입력해주세요"
-              autoFocus
               {...point}
             />
           </div>
-          <Button className="w-fit px-6">참여하기</Button>
+          <Button className="w-fit px-6">팟 생성 및 참여</Button>
         </form>
       </Modal>
     </>
